src/main/java/qna/domain/Answers.java
package qna.domain;

import qna.CannotDeleteException;

import java.time.LocalDateTime;
import java.util.ArrayList;
import java.util.List;

/**
 * Created : 2020-12-14 오전 9:59
 * Developer : Seo
 */
public class Answers {
    private final List<Answer> answers;

    public Answers(List<Answer> answers) {
        this.answers = answers;
    }

    public void add(Answer answer) {
        answers.add(answer);
    }

    public void isOwner(User loginUser) throws CannotDeleteException {
        for (Answer answer : answers) {
            answer.isOwner(loginUser);
        }
    }

    public List<DeleteHistory> delete() {
        List<DeleteHistory> deleteHistories = new ArrayList<>();
        for (Answer answer : answers) {
            answer.setDeleted(true);
            deleteHistories.add(new DeleteHistory(ContentType.ANSWER, answer.getId(), answer.getWriter(), LocalDateTime.now()));
        }
        return deleteHistories;
    }

    public int size() {
        return this.answers.size();
    }
}
